+++ src/main/java/io/github/astrapi69/crypt/data/algorithm/SignatureAlgorithmResolver.java
@@ -1,8 +1,33 @@
-package io.github.astrapi69.crypt.data.processor.certificate;
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.crypt.data.algorithm;
 
 import static io.github.astrapi69.crypt.data.factory.CertFactory.newX509Certificate;
 
 import java.lang.reflect.InvocationTargetException;
+import java.math.BigInteger;
 import java.security.KeyPair;
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
@@ -10,6 +35,7 @@ import java.security.PrivateKey;
 import java.security.PublicKey;
 import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
+import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -18,16 +44,18 @@ import java.util.TreeMap;
 import java.util.TreeSet;
 import java.util.logging.Level;
 
+import org.bouncycastle.asn1.x500.X500Name;
 import org.bouncycastle.cert.CertIOException;
 import org.bouncycastle.operator.OperatorCreationException;
 
-import io.github.astrapi69.crypt.data.algorithm.AlgorithmExtensions;
-import io.github.astrapi69.crypt.data.factory.CertificateTestDataFactory;
 import io.github.astrapi69.crypt.data.factory.KeyPairFactory;
 import io.github.astrapi69.crypt.data.key.KeyInfoExtensions;
 import io.github.astrapi69.crypt.data.key.KeySizeExtensions;
 import io.github.astrapi69.crypt.data.key.KeySizeInitializer;
 import io.github.astrapi69.crypt.data.model.CertificateInfo;
+import io.github.astrapi69.crypt.data.model.DistinguishedNameInfo;
+import io.github.astrapi69.crypt.data.model.ExtensionInfo;
+import io.github.astrapi69.crypt.data.model.Validity;
 import lombok.extern.java.Log;
 
 /**
@@ -36,7 +64,7 @@ import lombok.extern.java.Log;
  * signature algorithms can successfully generate an X.509 certificate
  */
 @Log
-public class CertificateVerifier
+public class SignatureAlgorithmResolver
 {
 
 	/**
@@ -101,10 +129,6 @@ public class CertificateVerifier
 				// Test each signature algorithm with the key pair
 				for (String signatureAlgorithm : signatureAlgorithms)
 				{
-					CertificateAlgorithmEntry certificateAlgorithmEntry = CertificateAlgorithmEntry
-						.builder().keyPairAlgorithm(keyPairAlgorithm)
-						.signatureAlgorithm(signatureAlgorithm).build();
-
 					CertificateInfo certificateInfo = CertificateInfo.builder()
 						.privateKeyInfo(KeyInfoExtensions.toKeyInfo(privateKey))
 						.publicKeyInfo(KeyInfoExtensions.toKeyInfo(publicKey))
@@ -170,6 +194,99 @@ public class CertificateVerifier
 			log.log(Level.WARNING, "Certificate creation failed for algorithm: "
 				+ certificateInfo.getSignatureAlgorithm(), e);
 			return false;
+		}
+	}
+
+	/**
+	 * Private class for generate dummy data for signature algorithm certificate verification
+	 */
+	private static class CertificateTestDataFactory
+	{
+
+		/**
+		 * Generates a new {@link X500Name} for the issuer with a common name of "Test Issuer"
+		 *
+		 * @return a new {@link X500Name} instance representing the issuer
+		 */
+		public static X500Name newIssuerX500Name()
+		{
+			return new X500Name("CN=Test Issuer");
+		}
+
+		/**
+		 * Generates a new {@link X500Name} for the subject with a common name of "Test Subject"
+		 *
+		 * @return a new {@link X500Name} instance representing the subject
+		 */
+		public static X500Name newSubjectX500Name()
+		{
+			return new X500Name("CN=Test Subject");
+		}
+
+		/**
+		 * Creates a new {@link DistinguishedNameInfo} for the issuer with predefined attributes
+		 *
+		 * @return a new {@link DistinguishedNameInfo} instance representing the issuer
+		 */
+		public static DistinguishedNameInfo newIssuerDistinguishedNameInfo()
+		{
+			String representableString = "C=US, ST=California, L=San Francisco, O=MyOrg, OU=MyUnit, CN=John Doe";
+			return DistinguishedNameInfo.toDistinguishedNameInfo(representableString);
+		}
+
+		/**
+		 * Creates a new {@link DistinguishedNameInfo} for the subject with predefined attributes
+		 *
+		 * @return a new {@link DistinguishedNameInfo} instance representing the subject
+		 */
+		public static DistinguishedNameInfo newSubjectDistinguishedNameInfo()
+		{
+			return DistinguishedNameInfo.builder().countryCode("GR").state("Pieria")
+				.location("Karitsa").organisation("Alpha Ro Group Ltd")
+				.organisationUnit("Certificate Authority").commonName("Alpha Robert").build();
+		}
+
+		/**
+		 * Creates a new {@link Validity} instance with the validity period starting from the
+		 * current date and time and lasting for one year
+		 *
+		 * @return a new {@link Validity} instance
+		 */
+		public static Validity newValidity()
+		{
+			ZonedDateTime now = ZonedDateTime.now();
+			return Validity.builder().notBefore(now).notAfter(now.plusYears(1)).build();
+		}
+
+		/**
+		 * Generates a new random {@link BigInteger} to be used as a serial number
+		 *
+		 * @return a random {@link BigInteger} instance
+		 */
+		public static BigInteger newSerialNumber()
+		{
+			return BigInteger.TEN;
+		}
+
+		/**
+		 * Generates a new {@link ExtensionInfo} for unit tests
+		 *
+		 * @return the created {@link ExtensionInfo} for unit tests
+		 */
+		public static ExtensionInfo newExtensionInfo()
+		{
+			return ExtensionInfo.builder().extensionId("1.2.3.4.5.6.7").critical(true)
+				.value("testValue").build();
+		}
+
+		/**
+		 * Generates a new array of {@link ExtensionInfo} objects for unit tests
+		 *
+		 * @return the created array of {@link ExtensionInfo} objects for unit tests
+		 */
+		public static ExtensionInfo[] newExtensionInfos()
+		{
+			return new ExtensionInfo[] { newExtensionInfo() };
 		}
 	}